requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.4.0
cached-property==1.5.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
ddt==1.4.4
decorator==5.1.1
Deprecated==1.2.13
et-xmlfile==1.1.0
facebook-wda==1.4.6
Faker==9.3.1
future==0.18.2
idna==2.10
iniconfig==1.1.1
jsonpath==0.82
openpyxl==3.0.9
packaging==21.3
Pillow==9.0.1
pluggy==0.13.1
py==1.11.0
pyasn1==0.4.8
pymongo==4.0.2
PyMySQL==1.0.2
pyparsing==3.0.7
pytest==6.2.3
pytest-html==3.1.1
pytest-metadata==1.11.0
python-dateutil==2.8.2
PyYAML==6.0
redis==4.1.4
requests==2.25.1
retry==0.9.2
rsa==4.8
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
urllib3==1.26.8
websocket-client==1.3.1
wrapt==1.13.3
